perf: optimize transforms of chartGenerator
