+++ packages/vmind/src/atom/chartGenerator/const.ts
@@ -24,9 +24,12 @@ import { ChartType, MapRegionCoordinate } from '../../types';
 
 export const SUPPORTED_CHART_LIST = Object.values(ChartType);
 
+/**
+ * 需要处理color 映射的图表
+ */
 export const NEED_COLOR_FIELD_CHART_LIST = [ChartType.PieChart, ChartType.RoseChart, ChartType.LinearProgress];
 
-export const NEED_SIZE_FIELD_CHART_LIST = [ChartType.ScatterPlot, ChartType.BasicHeatMap];
+export const NEED_SIZE_FIELD_CHART_LIST = [ChartType.ScatterPlot, ChartType.BasicHeatMap, ChartType.LiquidChart];
 
 export const NEED_VALUE_FIELD = [ChartType.SankeyChart, ChartType.CircularProgress];
 
@@ -38,8 +41,7 @@ export const NEED_COLOR_AND_SIZE_CHART_LIST = [
   ChartType.Gauge,
   ChartType.SunburstChart,
   ChartType.TreemapChart,
-  ChartType.CircularProgress,
-  ChartType.LiquidChart
+  ChartType.CircularProgress
 ];
 
 export const CARTESIAN_CHART_LIST = [
@@ -52,7 +54,8 @@ export const CARTESIAN_CHART_LIST = [
   ChartType.WaterFallChart,
   ChartType.BoxPlot,
   ChartType.RangeColumnChart,
-  ChartType.LinearProgress
+  ChartType.LinearProgress,
+  ChartType.BasicHeatMap
 ];
 
 export const DEFAULT_MAP_OPTION: BasemapOption = {
