+++ packages/vmind/src/atom/chartGenerator/spec/transformers/cartesian.ts
@@ -1,0 +1,91 @@
+import { isArray, uniqArray } from '@visactor/vutils';
+import type { GenerateChartCellContext } from '../../type';
+import type { DataCell } from '../../../../types/base';
+import { ROLE } from '../../../../types/base';
+
+export const seriesField = (context: GenerateChartCellContext) => {
+  const { spec, fieldInfo, dataTable, cell } = context;
+  const cellNew: any = { ...cell };
+  const { seriesField, xField: propsXField } = spec;
+  const colorField = isArray(seriesField) ? seriesField[0] : seriesField;
+  const colorFieldInfo = fieldInfo.find(v => v.fieldName === colorField);
+  const xField = isArray(propsXField) ? propsXField : [propsXField];
+  if (colorField && colorFieldInfo?.role === ROLE.DIMENSION && xField) {
+    const xMap = new Map<DataCell, DataCell[]>();
+    dataTable.forEach(row => {
+      const xValue = row[xField[0]];
+      if (xMap.has(xValue)) {
+        xMap.get(xValue).push(row[colorField]);
+      } else {
+        xMap.set(xValue, [row[colorField]]);
+      }
+    });
+    const xValues = Array.from(xMap.keys());
+    let isValidColor = false;
+    for (let i = 0; i < xValues.length; i++) {
+      const xValue = xValues[i];
+      const colorValues = uniqArray(xMap.get(xValue));
+      if (isArray(colorValues) && colorValues.length > 1) {
+        isValidColor = true;
+        break;
+      }
+    }
+    if (!isValidColor) {
+      spec.seriesField = undefined;
+      spec.xField = xField.filter((field: string) => field !== colorField);
+      cellNew.color = undefined;
+    }
+  }
+  return { spec, cell: cellNew };
+};
+
+export const axis = (context: GenerateChartCellContext) => {
+  const { spec, cell, fieldInfo } = context;
+  const { y: celly } = cell;
+  const yFields = isArray(celly) ? celly : [celly];
+  const yFieldsInfo = yFields.map(field => fieldInfo.find(v => v.fieldName === field));
+  const isAllRatio = yFieldsInfo.every(v => !!v?.ratioGranularity);
+  const isSameUnit = uniqArray(yFieldsInfo.map(v => v?.unit).filter(v => !!v)).length === 1;
+
+  spec.axes = [
+    {
+      orient: 'bottom',
+      type: 'band',
+      label: {
+        style: {
+          //fill: '#FFFFFF'
+        }
+      },
+      title: {
+        visible: false,
+        style: {
+          //fill: '#FFFFFF'
+        }
+      }
+    },
+    {
+      orient: 'left',
+      type: 'linear',
+      label: {
+        style: {
+          //fill: '#FFFFFF'
+        },
+        formatter: isAllRatio ? `{label:~%}` : undefined
+      },
+      unit:
+        isSameUnit && !['%', '‰'].includes(yFieldsInfo[0]?.unit)
+          ? {
+              visible: true,
+              text: yFieldsInfo[0]?.unit
+            }
+          : undefined,
+      title: {
+        visible: false,
+        style: {
+          //fill: '#FFFFFF'
+        }
+      }
+    }
+  ];
+  return { spec };
+};
